refactor(session): extract the proxy launcher into src/session/launch/proxy-launcher.ts
startProxyManager and its private helpers (prepareLaunchInputs,
prepareAdapterLaunch, transformAdapterConfig, recordAttachKeyDiff,
resolveAdapterExecutable, buildAdapterLaunchPlan, applyToolchainValidation)
move verbatim onto ProxyLauncher.start(), keeping the seven positional
arguments and the return type. The three launch intermediates
(ProxyLaunchRequest, LaunchInputs, AdapterLaunchPlan) move with it and are
exported; prepareLaunchInputs and buildAdapterLaunchPlan are public so they can
be unit-tested directly. The facade reaches the launcher through a new
ProxyLaunchContext slice of OperationsContext (late-bound, like the rest) and
calls this.proxyLauncher.start(...) at call time; startProxyManager is deleted
from the facade.

Test migration: every `(ops as any).startProxyManager` site (48 in the
coverage test, 5 in attach-modes) becomes `internals(ops).proxyLauncher.start`.
The view type in each test names the two loosenings the `as any` used to hide
(void-resolving spies, attach-shaped dapLaunchArgs) so the test-typing ratchet
stays at its baseline. Comments that named startProxyManager are updated; the
`Entering startProxyManager` log line is deliberately unchanged.
